backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
stripe==7.8.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
boto3==1.34.0
slowapi==0.1.9
redis==5.0.1
celery==5.3.4
pillow==10.1.0
python-magic==0.4.27 